More helpful error messages
The current error messages can be a bit confusing; especially the `cannot read property 'replace' of undefined` error (see #72).
This adds 2 more checks to give more helpful messages and replaces the `throw "<msg>"` in base64_url_decode with `throw new Error("<msg>")` as the former lead to undefined error messages.
Tests are added and the README is updated with the different error messages.